Cleanup util naming
