+++ mobile/src/components/account-item.js
@@ -13,7 +13,7 @@ import { connect } from 'react-redux'
 import { fbt } from 'fbt-runtime'
 import get from 'lodash.get'
 
-import { truncate } from 'utils/user'
+import { truncate } from 'utils'
 import Address from 'components/address'
 import Avatar from 'components/avatar'
 import ListStyles from 'styles/list'
