+++ mobile/src/utils/exchangeRate.js
